Put some order in the code and removed the dom actions from the index.js
